Project Development Phase/Sprint 1/app/requirements.txt
flask
ibm_db
pyjwt

sib-api-v3-sdk